Workd on Scanner Screen and, add and show Employee Details showing screens in Company Flow
