+++ src/Screens/CompanyFlow/ReceiptList/ReceiptListStyle.js
@@ -1,0 +1,160 @@
+import {StyleSheet} from 'react-native';
+import {SetFontSize} from '../../../Utils/FontSize';
+import {colors} from '../../../Utils/colors';
+import {ConstantKeys} from '../../../Utils/ConstantKey';
+import {
+  widthPercentageToDP as wp,
+  heightPercentageToDP as hp,
+} from 'react-native-responsive-screen';
+
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#FDFDFD',
+  },
+  headerView: {
+    height: hp(4),
+    width: wp(100),
+    backgroundColor: '#3461A6',
+    justifyContent: 'center',
+    paddingLeft: wp(7),
+    marginTop: hp(4),
+  },
+  headerTxt: {
+    fontSize: SetFontSize.setDimension.textSizeMedium,
+    color: colors.WHITE,
+    fontWeight: '600',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  mainView: {
+    borderWidth: 1,
+    backgroundColor: colors.WHITE,
+    borderColor: '#E8E5E5',
+    borderRadius: 10,
+    marginHorizontal: wp(2),
+    marginVertical: hp(2.5),
+    paddingBottom: hp(3),
+  },
+  topButtonView: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    marginTop: hp(2.5),
+    marginHorizontal: wp(4),
+  },
+  budgetButtonView: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    height: hp(3.5),
+    width: wp(21),
+    borderWidth: 1,
+    borderColor: '#C0C0C0',
+    borderRadius: 5,
+  },
+  exportButtonView: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    height: hp(3.5),
+    width: wp(21),
+    borderWidth: 1,
+    borderColor: '#C0C0C0',
+    borderRadius: 5,
+    marginLeft: wp(15),
+  },
+  searchButtonView: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    borderWidth: 1,
+    borderColor: '#C0C0C0',
+    borderRadius: 5,
+    height: hp(3.5),
+    width: wp(21),
+  },
+  buttonImg: {
+    marginTop: hp(0.3),
+  },
+  upperButtonTxt: {
+    fontSize: SetFontSize.setDimension.textSize13,
+    color: colors.BLACK,
+    fontWeight: '400',
+    fontFamily: ConstantKeys.POPPINS,
+  },
+  bodyView: {
+    flexDirection: 'row',
+    marginTop: hp(5),
+    backgroundColor: '#F4F4F4',
+    justifyContent: 'space-between',
+    marginHorizontal: wp(1),
+  },
+  categortTxt: {
+    fontSize: SetFontSize.setDimension.textSize15,
+    color: colors.BLACK,
+    fontWeight: '500',
+    fontFamily: ConstantKeys.MUKTA,
+    marginLeft: wp(5),
+  },
+  totalTxt: {
+    fontSize: SetFontSize.setDimension.textSize15,
+    color: colors.BLACK,
+    fontWeight: '500',
+    fontFamily: ConstantKeys.MUKTA,
+    marginLeft: wp(10),
+  },
+  receiptTxt: {
+    fontSize: SetFontSize.setDimension.textSize15,
+    color: colors.BLACK,
+    fontWeight: '500',
+    fontFamily: ConstantKeys.MUKTA,
+    marginRight: wp(5),
+  },
+  flatlistView: {
+    marginTop: hp(2),
+  },
+  listView: {
+    width: wp(200),
+    backgroundColor: colors.WHITE,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    shadowColor: colors.BLACK,
+    shadowOffset: {height: 1, width: 1},
+    shadowOpacity: 1,
+    elevation: 5,
+    shadowRadius: 1,
+    marginBottom: hp(1.8),
+    borderRadius: 10,
+    paddingHorizontal: wp(3),
+    paddingVertical: hp(0.5),
+    borderWidth: 1,
+    borderColor: '#EDEDED',
+    marginHorizontal: wp(1),
+  },
+  listLeftView: {flex: 1},
+  listRightView: {
+    flex: 1,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  listTxt: {
+    fontSize: SetFontSize.setDimension.textSize15,
+    color: '#4D4B4B',
+    fontWeight: '400',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  buttonView: {
+    marginTop: hp(0.2),
+    height: hp(3),
+    backgroundColor: '#3461A6',
+    width: wp(15),
+    alignItems: 'center',
+    justifyContent: 'center',
+    borderRadius: 5,
+  },
+  buttonTxt: {
+    fontSize: SetFontSize.setDimension.textSizeRegular,
+    color: colors.WHITE,
+    fontWeight: '500',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+});
